Files for initial webpage to setup the websocket
This is the very first trial run to get the websocket running and to initiate the communication between the webpage and the websocket. 
This will be further modified to improve the webpage and also to have proper communication established. This will also be modified to have the serial communication betweent the Raspberry Pi and the Arduino